bugfix: new receive add button in wallet widget was generating multiple address in 1 click
whenever a new coin added the script was executing the
Dashboard.init(), to refresh, and along with it it was also executing
functions multiple time, causing same functions to execute multiple
times, as the init was running as much time the Dashboard.init() was
running. this update should fix the issue.