2 Important Bugs fixed in send transaction screen
- if user hit on confirm button twice the same amount of transaction
goes twice!
Even while the transaction is being processed at the time!
It is fixed.

- Second bug I found while testing is if I send a transaction and it
completes. And I’m still on that same send transaction screen. I make
second transaction. When I confirm to send this transaction, not just
this transaction is processed to send, but the previous transaction as
well!!!
It is also fixed.